# src/main/java/com/example/POCBankService/model/AccountBalance.java
package com.example.POCBankService.model;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.math.BigDecimal;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class AccountBalance {
    private String accountId;
    private String accountName;
    private BigDecimal balance;
}