# src/main/resources/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/user/preregister":{"post":{"tags":["user-info-controller"],"operationId":"preRegister","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequestModel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/user/preregister/approval":{"post":{"tags":["user-info-controller"],"operationId":"approveOrRejectUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApprovalRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/user/customerregister":{"post":{"tags":["user-info-controller"],"operationId":"customerRegister","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRegisterRequestModel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/transactions/transfer":{"post":{"tags":["transaction-controller"],"operationId":"transfer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/transactions/deposit":{"post":{"tags":["transaction-controller"],"operationId":"deposit","parameters":[{"name":"account","in":"query","required":true,"schema":{"type":"string"}},{"name":"amount","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/user/preregister/list":{"get":{"tags":["user-info-controller"],"operationId":"listPendingPersons","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserInfo"}}}}}}}},"/api/user/me":{"get":{"tags":["user-info-controller"],"operationId":"getUserDetails","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDetailResponse"}}}}}}},"/api/transactions/statement":{"get":{"tags":["transaction-controller"],"operationId":"getStatement","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionStatementResponse"}}}}}}}},"/api/transactions/accounts":{"get":{"tags":["transaction-controller"],"operationId":"getAccounts","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountBalance"}}}}}}}}},"components":{"schemas":{"RegisterRequestModel":{"required":["citizenId","email","nameEN","nameTH","password","pin"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"citizenId":{"type":"string"},"nameTH":{"type":"string"},"nameEN":{"type":"string"},"pin":{"type":"string"}}},"ApprovalRequest":{"type":"object","properties":{"account":{"type":"string"},"action":{"type":"string"}}},"CustomerRegisterRequestModel":{"required":["nameEN","nameTH","userId"],"type":"object","properties":{"userId":{"type":"string"},"nameTH":{"type":"string"},"nameEN":{"type":"string"}}},"TransferRequest":{"type":"object","properties":{"userId":{"type":"string"},"pin":{"type":"string"},"sourceAccount":{"type":"string"},"destinationAccount":{"type":"string"},"amount":{"type":"number","format":"double"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"UserInfo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"thaiName":{"type":"string"},"engName":{"type":"string"},"email":{"type":"string"},"pin":{"type":"string"},"role":{"type":"string","enum":["PERSON","CUSTOMER","TALLER"]},"account":{"type":"string"},"userId":{"type":"string"},"password":{"type":"string"},"updateDate":{"type":"string","format":"date-time"},"updateBy":{"type":"string"},"createDate":{"type":"string","format":"date-time"},"createBy":{"type":"string"},"status":{"type":"string","enum":["PENDING_APPROVE","APPROVED","REJECTED","DELETED"]}}},"UserDetailResponse":{"type":"object","properties":{"userId":{"type":"string"},"role":{"type":"string"}}},"TransactionItem":{"type":"object","properties":{"date":{"type":"string"},"time":{"type":"string"},"code":{"type":"string"},"channel":{"type":"string"},"debitCredit":{"type":"string"},"balance":{"type":"number","format":"double"},"remark":{"type":"string"}}},"TransactionStatementResponse":{"type":"object","properties":{"account":{"type":"string"},"transactions":{"type":"array","items":{"$ref":"#/components/schemas/TransactionItem"}}}},"AccountBalance":{"type":"object","properties":{"accountId":{"type":"string"},"accountName":{"type":"string"},"balance":{"type":"number"}}}}}}